requirements.txt
# requirements.txt

blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
pycparser==2.22
Werkzeug==3.0.3
pandas==2.2.2
psycopg
python-dotenv
boto3
fsspec
geopandas
openpyxl
pillow
odfpy
py7zr
bcrypt
sqlalchemy
bcrypt
matplotlib
s3fs
pyexcel-ods3
